+++ cmd/frontend/internal/usagestats/saved_searches.go
@@ -1,0 +1,47 @@
+// Package usagestats provides an interface to update and access information about
+// individual and aggregate Sourcegraph users' activity levels.
+package usagestats
+
+import (
+	"context"
+
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
+	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
+)
+
+func GetSavedSearches(ctx context.Context) (*types.SavedSearches, error) {
+	const q = `
+	SELECT
+	  (SELECT COUNT(*) FROM saved_searches) AS totalSavedSearches,
+	  (SELECT COUNT(DISTINCT user_id) FROM saved_searches) AS uniqueUsers,
+	  COUNT(*) FILTER (WHERE event_logs.name = 'SavedSearchEmailNotificationSent') AS notificationsSent,
+	  COUNT(*) FILTER (WHERE event_logs.name = 'SavedSearchEmailClicked') AS notificationsClicked,
+	  COUNT(DISTINCT user_id) FILTER (WHERE event_logs.name = 'ViewSavedSearchListPage') AS uniqueUserPageViews
+	FROM
+	  event_logs
+	`
+	var (
+		totalSavedSearches     	int
+		uniqueUsers     	    int
+		notificationsSent 	 	int
+		notificationsClicked 	int
+		uniqueUserPageViews    	int
+	)
+	if err := dbconn.Global.QueryRowContext(ctx, q).Scan(
+		&totalSavedSearches,
+		&uniqueUsers,
+		&notificationsSent,
+		&notificationsClicked,
+		&uniqueUserPageViews,
+	); err != nil {
+		return nil, err
+	}
+
+	return &types.SavedSearches{
+		TotalSavedSearches:     int32(totalSavedSearches),
+		UniqueUsers:     	    int32(uniqueUsers),
+		NotificationsSent: 	    int32(notificationsSent),
+		NotificationsClicked:   int32(notificationsClicked),
+		UniqueUserPageViews:    int32(uniqueUserPageViews),
+	}, nil
+}