+++ cmd/frontend/types/types.go
@@ -390,6 +390,16 @@ type GrowthStatistics struct {
 	RetainedUsers    int32
 }
 
+// SavedSearches represents the total number of saved searches, users
+// using saved searches, and usage of saved searches.
+type SavedSearches struct {
+	totalSavedSearches     	int32
+	uniqueUsers     		int32
+	notificationsSent 		int32
+	notificationsClicked    int32
+	uniqueUserPageViews    	int32
+}
+
 // Secret represents the secrets table
 type Secret struct {
 	ID int32